moving identificationTypes type to dedicated file
