+++ src/coreMethods/getIdentificationTypes/types.ts
@@ -1,0 +1,7 @@
+export type IdentificationType = {
+  id: string,
+  name: string,
+  type: string,
+  min_length: number,
+  max_length: number
+}
