+++ src/mercadoPago/initMercadoPago/type.ts
@@ -1,3 +1,4 @@
+import type { IdentificationType } from "../../coreMethods/getIdentificationTypes/types";
 import type { TPaymentMethods, TPaymentMethodsParams } from "../../coreMethods/getPaymentMethods/types";
 
 export type TOptions = {
@@ -21,17 +22,6 @@ export type TOptions = {
 export type BricksBuilderType = {
   create: (param: string, param2: string, settings: {}) => void;
 };
-
-// #############
-// CORE METHODS
-// #############
-export type IdentificationType = {
-  id: string,
-  name: string,
-  type: string,
-  min_length: number,
-  max_length: number
-}
 
 export type TInstanceMercadoPago = {
   bricks: () => BricksBuilderType;
